Fix Exception Handling and Non-Decorated Methods
In some cases exceptions thrown inside decorated code would be wrapped
in another exception. This was fixed and now the original, unmodified
exception is thrown to the call site.

Non decorated methods in a decorated class would throw an index out of
range exception. This was due to invalid logic in check whether the
method is decorated.