Add explicit ref to System.Net.Http >=4.3.4
Older System.Net.Http versions have a vulnerability (see
https://github.com/advisories/GHSA-7jgj-8wvc-jh57).

The package System.Net.Http is not a direct requirement of this package,
but it is a transitive dependency coming from Castle.Core.

Castle.Core have adressed this issues in >=5.0.0, but they did not
release a fix for versions <5.0.0. Bumping a dependencie's major version
would be a breaking change.

So instead, this is an attempt to circumvent the vulnerability by
specifying that vulnerable versions are not compatible with this
package.